requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-daraja==1.3.0
drf-yasg==1.21.11
gunicorn==21.2.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
pytz==2025.2
sqlparse==0.5.5